don't submit a 'reboot' command
This was actually just a mistake in commit b4923f03, the else branch
should have been there to begin with.

The tests don't pick this up because it's *also* passing 'off' and 'on'
so it does actually work correctly, it just triggers an extra error
message about the unrecognised 'reboot' command.

Change-Id: I31602975822f8c888628f4bde6bce06d38c1d272