use absolute TFTP paths in GRUB2 configs
GRUB2 just uses the filenames literally as they are given in the config
file, it doesn't resolve them relative to the config path or anything.
So currently it is sending TFTP requests for paths like:

    ../images/system.example.com/kernel

As it happens, the TFTP server just strips off the initial .. and
honours the request, which is why it has always seemed to work. But
really it should just be requesting an absolute TFTP path starting
/images/.

Change-Id: Ic98be66af49aac21a428353538c10305ead36d86