explicitly index map tables' first columns
In old databases, the various association tables will have an index on
their first column. It is a left-over implicit index created by MySQL
because they had no primary key. In newly created databases this is
absent because MySQL uses the PK instead of an implicit index.

This is a side-effect of commit ca7c56e6 for bug 642834.

We can just make the indexes explicit, and add them on newer databases
where they are not already present, because why not.

Change-Id: Ib7794141b94fadcdbfa4844a2f4c280bf8f2f272