beaker-init: need to stamp revision after determining it
This was a mistake in commit 2df6c707, the patch for determining the
schema version for an unstamped database.

The version determination worked properly but the version was never
actually stamped, so the upgrade would proceed but alembic_version would
be left empty. Re-running the upgrade again would see an unstamped
database which would then repeat the post-0.18 upgrades, which would
then fail.

Bug: 682030
Change-Id: Ic3bc8acd7c1772e2d195b7a54a0d404fdb8acb13