only show scary warning on Provision tab if a recipe is running
Instead of always showing a big fat "Caution", which just causes extra
visual noise, we show a shorter and less obtrusive message only when the
system is running a recipe.

That is the only case where the behaviour in Beaker 19 will be different
than earlier versions (previously it would implicitly schedule a job
instead of provisioning) and it's also the only time where provisioning
is probably not what the user wants to do.

Also updated the confirmation box to match the wording from the power
confirmation box, in case the system is reserved by someone else.

Change-Id: I5aa4812da0118716f1c4f72a7248593e9927b445