fix reservesys notification for OpenStack instances
Currently Beaker is sending out a link to /view/<fqdn> which is not
correct for OpenStack instances.

Converted to use a Jinja template since we now have some logic in there.
Also updated the tests to use more known values so that we have less of
the original template logic repeated in the tests.

Change-Id: Idf5028df410db689f641c33c6c448986af38f0ae