pass -noreset to Xvfb
This option tells Xvfb not to "reset" itself after the last client
disconnects (which happens every time Firefox is killed, since it is the
only client).

While Xvfb is resetting there is apparently a window during which it is
not accepting connections. This is the cause of intermittent "cannot
open display" errors from Firefox, if it starts up while Xvfb is in the
middle of a reset.

Many thanks to Peter Hutterer for pointing this out!

Change-Id: I6ea1e8a82f90130a2414e17679201422f7b29aee