fix timing issues with logout/login
Newer Firefox+Selenium revealed a timing issue with the logout()
function. It does not wait for the page load to finish after logging
out, so a test may send another request straight away which could be
processed before the logout was actually handled.

The particular case I saw this was with logout() immediately followed by
login(). Since login() re-loads / it may see a page that is still logged
in from before.

I also removed the extra step of loading / at the start of login(), to
avoid some unnecessary roundtrips, although in some cases we still need
to do it because the test is calling login() before it loads any page at
all.

Change-Id: Ib730387662493eaf0aea22ffe371f083a5b81000