tidy up usages of /tmp in tests
Removed all hardcoded /tmp paths. There are a few things (like
basepath.harness) which we can't put in a dynamically created /tmp
directory, so they are moved to ./ when running tests from a git
checkout (with a corresponding .gitignore rule).

Also ensured that all temp files and directories are cleaned up. The
only gotcha is that the cleanup does not always happen if nose is
interrupted during a run.

The end result is that for a fully passing test run, nothing will be
left behind in /tmp.

Change-Id: Iff060ea5b45ca028af128fce98ac4e3b4b40782e