fix population of command grid when submitting a new command
Because the System#command_queue attribute is actually
a PageableCollection running in server mode, it could be currently
storing any page -- possibly with a sort or filter applied. So we can't
just always insert the new command at the top and assume that is the
right place.

The easiest approach is to just refresh the command queue grid from the
server after we have finished enqueuing our new command(s).

This patch also updates the 'reboot' logic so that it first saves 'off'
and then only saves 'on' after that has succeeded, to ensure they are
ordered correctly.

Bug: 1144193
Change-Id: Ied0e3744ebc30be94e7741bcaedb80169876f45d