Misc/rpmbuild.sh: match all tags, not just beaker-0*
The beaker-0* pattern was needed to avoid matching task tags (such as
beaker-distribution-install-1_11-11 etc). Now that we have dropped the
leading zero from Beaker versions that pattern is no longer suitable.
Since we are not keeping tasks in Beaker's git repo anymore we can just
remove the pattern entirely and match all tags.

Change-Id: I7579e83b710727e27ad12d988eb536a70f89e2fd