tests: show complete diff in assertion message for kickstart tests
The diffs are really long and often noisy but I don't care. It's still
easier than trying to debug a failure in dogfood where you have no
access to the tmp files that are left behind by the failures.

Not leaking tmp files on failure is really a bonus as well.

Change-Id: If25b29358869f9e071182356b892268c933ad8d9