beaker-init: split db population step out, always run it
This actually changes it back to match the pre-19.0 behaviour: the
schema is first created (or upgraded) if necessary, and then any
skeleton values are populated if necessary.

With this change the logic for setting up new admin users, or converting
existing users into admins, will always take effect no matter whether
you are starting from an empty schema, an outdated schema, or an
up-to-date schema.

Bug: 745560
Change-Id: I997a851d620ee6d391b942c32b3e1ea402006361