docs: flesh out docs for kickstart templates
Added a section to the user docs with complete details of all the
variables and constructs available to custom kickstart templates.

Changed the admin kickstart template section to reference the user
section, plus a description of the extra variables available for admins.

ks= and ks_appends are also now mentioned in the user doc about
customising the installation.

Bug: 966348
Change-Id: I1e55436a2555a72bce9ac3214584ab5b7cab6fe7