avoid ORDER BY when counting rows in @json_collection
The query from the decorated function is expected to have a default
ordering pre-defined, which is fine, except when we issue the SELECT
COUNT(*) query we need to avoid including the ORDER BY because MySQL is
not smart enough to skip the ordering step when counting rows.

So when the default ordering is used we can save a lot of query time by
skipping the ORDER BY when issuing the count.

Change-Id: I9332e2e1d9a53fea190f280a83b83e069d7bb422