system search bar: handle invalid date
For the search fields which are using jQuery UI Datepicker, if a user
manually types an invalid date, it will be either handled by the date picker
or HTML5 form validation.The date picker will restore the value to default
current date. Also, if the user puts in a date which can be interpreted,
such as "15-05-01", it will reformat the date as "2015-05-01".

Bug: 1182545
Change-Id: I05dceca62a9f1bb5e8fe6ace7cd8f5b14ff444bb