BeakerGrid: don't fetch in initialize
The code currently assumes that the collection is not bootstrapped into
the page, which is true for all the grids we have now on the system page
(they are fetched on demand). However it is not true for grids we will
be adding in future. So the BeakerGrid widget no longer triggers any
fetches itself, it's up to the caller to do that if necessary.

Change-Id: If460c5ae9270825fed3191cf41f1775e98c54ba7