skip COUNT(*) queries for activity because they are too expensive
The activity table on a large Beaker installation is typically very
large (tens of millions of rows) and so counting all rows in an activity
query is expensive, and can in some cases results in a temp table
overflowing disk space.

For activity types which are likely to accumulate many rows, we skip the
count. We have to just leave out the 'count' key from the response in
that case.

Updated the Backgrid paginator to handle an unknown total row count.

Bug: 1101817
Change-Id: I632435f52176dbc3fefd6fdbf1634bc64594c458