fix very slow activity queries
We are fighting a losing battle with this stuff... but adding this
IGNORE INDEX seems to convince InnoDB to select from the relevant
*_activity table first, rather than the associated table (system, etc),
thus avoiding a filesort that fills up the disk.

All the type-specific activity queries suffer this problem although
system activity is the only one where the size of the temp table is
likely to be large enough to cause a problem.

Change-Id: I4041ac4bdc9bc8310cb1c7110dbe90458c180cae