tests: add db migration test from 19, matching the others
Updated the migration tests so that we have a test for migration from
the Beaker 19 schema, matching all the other similar tests.

The tests covering pool and policy migration don't need to also check
the upgraded schema, we just do that in one place instead.

Also re-generated the 19.sql db dump in the same way as the others, so
that it contains the bare minimum data left by beaker-init, and in
a form normally produced by mysql_dump. (The bare INSERT with no
surrounding transaction in the previous 19.sql file was causing the test
to hang with a lock wait timeout.)

Change-Id: Ie99522ae047e4459aadf6a606756b0a744144e13