fix grid loading indicators
This was a mistake in commit 21f2418a. The tests didn't catch it then
(we don't test the loading indicators directly, because we can't really)
but it did cause intermittent timing-sensitive failures in the grid
tests because they were no longer properly waiting for the grid to load.
So this patch should fix those problems too.

Change-Id: I25b3df439409d7285cb0635da7ad346f2d632eea