Misc/rpmbuild.sh: --next-major option to increment major version
Passing --next-major as the first option to rpmbuild.sh will cause a git
build to appear as a pre-release for the next major version (rather than
the next minor version, as it does currently).

We can set this in the Jenkins job for nightly develop RPMs, to
distinguish them better from the maintenance branch.

Change-Id: I49cd04686086062958a2e03bdd841c46162959c0