increase requested size of /boot to match documented recommendations
On RHEL7 and newer we can omit the size and just pass --recommended and
Anaconda will pick the right size (currently 500MB). For older releases
we still need to supply a fixed size.

The sizes in this patch are based on the recommendations in the RHEL
product documentation. The related bug has links.

Bug: 1262098
Change-Id: I400bab5a77ac78d1a88409f2c9114719b59fa441