Fix marking systems as broken
This patch fixes the detection of possible broken systems. Release 21
introduced a change as to when Beaker regards systems to be broken. By
just looking at the Recipe status, systems were already regarded as
broken if any of the tasks in a Recipe were set to aborted. This patch
labels systems as broken if *all* tasks in a Recipe are set to aborted.

The patch uses a hybrid method on a recipe to ensure that we call
similar functionality from two different points in the code:

        1. During an update status, we only want to trigger the
        suspicious_abort if all tasks in the current recipe are set to
        aborted.
        2. Once (1.) is satisified, we only want to mark a system as
        broken, if - for the past two recipes - all tasks are aborted.

Bug: 1270649
Change-Id: I66955ec7a19bd339e3d31cb81ad745a9d982a01c