add SQLite support for ConditionalInsert
Previously this hardcoded some MySQL-isms, which meant it didn't work in
SQLite. As a consequence, any code calling lazy_create(), which is
almost everything, could not be used in unit tests running against
SQLite and had to be in IntegrationTests so that it ran against MySQL.

This lets us write more unit tests that don't need an entire MySQL
server to run.

Change-Id: Id04bbefbacf92e77400ebbcc77b2828d75125b25