don't run unit tests with integration tests
For historical reasons we have always just run all tests as part of the
IntegrationTests/run-tests.sh script, since there wasn't really anywhere
better to run them. However we now have separate scripts (and make
targets) for running unit tests in each of the source trees, so we don't
need to run them here too.

The reason to avoid running unit tests as part of the integration test
suite is that it's sometimes difficult to write unit tests which work
properly as both a regular unit test and also with the integration test
config, which is substantially different. I discovered this when working
on unit tests for bkr.server.identity in bug 1112925.

(Also dropped rhts from the nose invocation because it doesn't need to
be there anymore, we stopped using any rhts modules in Beaker ages ago
in commit 3348ae74.)

Change-Id: I866e018d50d154f7271a7af33b15ec56cc183412