Misc/rpmbuild.sh: ensure NVR for develop does not decrease when a minor release is tagged
Currently the RPM release has 0.git.${count}.${sha} where the count is
the number of commits since the most recent tag. However that means when
we tag a minor release (e.g. 21.1) and merge it into develop (e.g.
22.0-0.git...), the commit count *decreases*, even though the RPM
release should continue to *increase*.

We can avoid that by counting commits from the previous major release
instead of just the nearest tag, so that we get 22.0-0.git.58.22f027d
which is 58 commits from 21.0.