fix primary key on config value tables
The config_value_string and config_value_int tables have a composite
primary key across (id, config_item_id). Not for any particular reason,
since id is already autoincrementing. Probably it was originally
intended to be (config_item_id, valid_from) or similar.

The composite PK does not in itself cause any problems, but in SQLite
for unit tests it causes the id autoincrementing not to work. So this
patch changes the PK to be just on (id), which has equivalent behaviour
but lets unit tests create config values.

Change-Id: Ief1bfb3d9c1d21552ba9aab7e2072b8fc4ea20d9