refactored RecipeTaskResult.short_path to just return a path
Introduced a new property display_label which preserves the previous
behaviour of short_path, where the log is shown if the path is empty,
which is really a display issue specific to the tasks widget.

This is needed so that we can use short_path in JUnit XML to get just
a path.

Bug: 1123244
Change-Id: Id875daa7e68970d856635b83906d7e7e2118c278