Conduct coverage analysis when running tests
This allows measuring test coverage of code run in a separate process,
the server in particular.

Two challenges with enabling coverage analysis:

    1) multiple worker processes
    2) coverage analysis conducted by the test runner

The coverage module allows measuring coverage in separate processes.
Data files are written by every process. It is then up to the test
runner to combine all data files and report the coverage metric.

Change-Id: Id704eeeba2e832a91f6484077d0a1617ee3b0866