Fix cause of memory leak in tests
This fixes the cause of a memory leak when running wizard test cases.
`mkdtemp` created temporary directories with an underscore which the
wizard detected as invalid package names. To avoid that, this fix
creates a sub directory inside the temporary directory which is always
valid to avoid the wizard ending up in an infinite loop waiting for
input.

Change-Id: I30c394c4381bc88c8c3ba41726b6010853b15332