JUnit XML: don't report duration for overall task
In the JUnit XML results we include a <testcase/> for the overall task
(in addition to all its results) so that we have somewhere to put the
task logs. However we *don't* want to include a duration for the overall
task because that will be summed with the individual result durations
making the task look like it took twice as long as it really did.

Bug: 1291107
Change-Id: If58689c543b9aab68b78cbc660e43f20ed5389f6