tests: ensure recipe's distro tree is available in the lab where it is running
An important invariant enforced by the Beaker scheduler is that a recipe
set always runs on systems attached to a lab where the requested distros
are also available. We need the tests to ensure the same.

The reason this matters is because currently in the tests a given
running recipe might use a distro which is or isn't available in the
lab, depending on prior test runs.

This patch fixes data_setup to ensure the distro tree is available in
the lab, and also fixes a number of existing tests which were breaking
this invariant.

Change-Id: Idf13ad60353cd3e01f47221c0ea5131f7512d6f8