beaker-init --background
New option to run detached from the terminal. Includes some pid file
handling to make it possible for other tools to wait for the background
beaker-init process to complete.

Also expanded on the log messages in beaker-init, to make it easier to
tell what has happened when it is backgrounded and logging to syslog.

Bug: 1302942
Change-Id: I036ac7b56e9bca2258022880e92566ab3fd14561