strip surrounding whitespace from <whiteboard/> in job XML
This restores the pre-22.0 behaviour, where xmltramp was implicitly
stripping the text content when we call unicode() on the element.

We need to keep this behaviour because some job XML generators
(including all existing versions of the bkr workflow commands) inject
extraneous whitespace around the whiteboard assuming that the server
will strip it out.

Bug: 1302857
Change-Id: I6deb97c3c4dec64938cc172637d8d0bcba1cdae9