redirect unauthenticated humans to the login form
We still want to serve text/plain error messages to API requests and
other scripts, but for humans looking for HTML we can redirect them to
the login form the way that the TurboGears controllers currently do.

This has become necessary in recent Beaker versions now that we have
a number of human-visible pages being served by Flask directly without
TurboGears.

Change-Id: I9aab266b8490923c9658601da0ac53a685dad4aa