add installation table
This new table consolidates information about an installation done on
a system, which is currently spread amongst the recipe_resource and
command_queue tables.

The goal is to make it possible to go from a recipe to the associated
work that was done to install it (including generated kickstart, kernel
options, and power commands).

A future patch could make the list of installations done on a system
available in the UI.

Bug: 991245
Change-Id: Ib4fcd6bd4e2268097ac0f9b3b803d772801e1bd9