reduce memory usage in beaker-log-delete
This patch updates beaker-log-delete to use Job.all_logs() for fetching
all logs in a job, rather than traversing the entire hierarchy of
objects down to RecipeTaskResult.

The Job.all_logs property was introduced in commit 377ccaa6 for reducing
memory usage in some XMLRPC handlers. The idea here is the same, we want
to avoid loading all RecipeTask and RecipeTaskResult instances into
memory because on large jobs that would mean a lot of database
roundtrips and a lot of memory consumed.

It was necessary to convert Job.all_logs() into a method, so that the
caller can opt out of having .parent populated by contains_eager() --
since that's exactly what we are trying to avoid in this patch.

Based on patches by Roman Joost <rjoost@redhat.com> and
Matt Jia <mjia@redhat.com>.

Bug: 1293011
Change-Id: I6fe983475ad1b5cb2b7ac5ecc379a3ef40de300a