recipe page: reflect expanded task rows in the URL anchor
Currently the page uses the link_tabs_to_anchor() function which gives
a rudimentary mapping from a set of tabs to the URL anchor. But for the
recipe page we have an additional complexity: we *also* want to
represent which tasks are expanded in the anchor.

So #tasks will correspond to the Tasks tab with nothing expanded,
#task123 will correspond to the Tasks tab with task 123 expanded, and
#task123,task234 similarly with both task 123 and task 234 expanded.

This patch introduces a new Backbone view for managing the entire layout
of the recipe page, instead of having that hardcoded in recipe.kid. This
new view is also responsible for keeping the view state (that is, which
tab is displayed, and which tasks are expanded) in sync with the URL
anchor.

Change-Id: Ie1804ea663e79dc71531407accd4ed6be2138395