simplify client login logic (no logout or renew_session)
The client code has no support for persisting sessions/cookies to disk,
so the call to auth.renew_session is pointless (the answer will always
be, yes you need to renew your session) and any calls to auth.logout are
also pointless (the session is discarded as soon as the process ends).

I suspect the original code from Kobo was careful about not leaking
client sessions because they cost a database row on the server side.
However since Beaker 0.15 we are using signed tokens which have no
storage cost on the server side, so that concern is no longer relevant.

This also saves two roundtrips to the server on every bkr invocation
which will be nice for high latency clients.

Change-Id: Ie25b4f01b71cca9e8b64d6e5c2a03c9bf833bef6