separate "online data migration" step for populating commands <-> installations
This data migration step, for matching up old command_queue rows with
their corresponding recipe installation row, is extremely expensive
(over 24 hours for our production database) so we don't want to do it
while Beaker is in outage. We can safely do it while Beaker is running,
after the actual schema upgrades are done.

So we introduce a new option, beaker-init --online-data-migration, which
will perform the named online migration step after the schema upgrade is
complete. In future when we have similar migration steps we can add
a new script which can be invoked in the same way.

Bug: 991245
Change-Id: I9fa61c3fa6c4d5d1a9b4b7acd1ef803503f9f9fc