fix installation.created for guest_resource rows
We set it to the start time of the reservation for the host recipe,
which is really when the guest recipe provisioning was done.

We still a few thousand older rows where there is actually no
recipe_resource for the host recipe at all, in which case there is no
possible timestamp we can use -- just fall back to the epoch for those.

Bug: 991245
Change-Id: I88d38a5254867d861d5a00623d1dcd0cde0738f9