tidy up usage of bytes/unicode from lxml.etree.tostring()
In some places, we really do want bytes so this patch just makes the
encoding explicit in those cases.

In some other places, where the serialized XML is stored back into the
database, we actually want unicode not bytes. This way we avoid an extra
UTF-8 conversion round-trip, as well as avoiding warnings like this one:

    /usr/lib64/python2.6/site-packages/sqlalchemy/engine/default.py:459: SAWarning: Unicode type received non-unicode bind param value
      param.append(processors[key](compiled_params[key]))

And in lots of test code we were mixing bytes and unicode, which works
because our test data was plain ASCII but is not strictly correct. And
one day we would like to be Python-3-ready.

Change-Id: Ie320f37ab5aab0b1c3ab99c39073c6c8753e3739