mark job as dirty whenever installation timestamps change
The recipe status can change depending on which timestamps are present
on the associated installation row, so we need to ensure the job is
marked dirty whenever the timestamps change so that beakerd will pick it
up and call update_status on the job.

Use a SQLAlchemy 'set' event listener instead of trying to modify all
calling sites to make the appropriate call (which is what I had here
before, with the recipe.installing() method).

Also added some invariant checking to job.update_status() to try and
catch cases like this in future, where the tests are assuming that
beakerd will pick up the job for update_status but it's not actually
dirty.

Bug: 991245
Change-Id: Ia2a765c2156535f0ee4bb7664868ceb3a0b0a0c4