ipxe-image: make the image be a whole number of tracks
Starting from dosfstools 4.0, mkdosfs is producing different C/H/S
geometry which no longer results in a whole number of tracks in the
filesystem. That causes mcopy to complain.

The actual geometry values used are meaningless but we do need mcopy to
be happy, so we switch to using mformat instead of mkdosfs, so that we
can pass explicit geometry when creating the filesystem.

The numbers used here (4, 64, 32) are completely arbitrary but they
match the geometry which dosfstools 3.x was creating, and they produce
a 4MB filesystem with a whole number of tracks which keeps mcopy happy.

Change-Id: I7e792c99ba6ed1518bc6cf3387fa8c13490457c5