right-align comment links on recipe page
This was done on the job page already, update the recipe page to match.

Unfortunately we also have to expand the right-most flex box for status,
otherwise the comment link (which now sits a bit further to the right,
due to the right alignment) will cause the comment popover to be placed
with its right edge partially outside the viewport.

Change-Id: Ica33c5b70959ee90a18b74c048884e73a463d417