fix URL generation for Job model
This is needed on the recipe page, where the Job instance is nested
inside the RecipeSet (which is inside the Recipe) so the .url property
is not set.

Currently when the Job model on the recipe page changes, it's triggering
an error when it tries to refresh JobActivity but there is no .url set.

Change-Id: I4736d6647d0fb09ef01bfe21f78ec5f21e53ca2d