don't trim recipe duration
We want {trim: false} here, otherwise it's possible to end up with
a summary sentence like:

    Started 2 minutes ago and finished in 01.

where the "01" means 1 second. Instead we want:

    Started 2 minutes ago and finished in 00:00:01.

which is unambiguous.

That's obviously a pathological case (I observed it in a test case,
a normal recipe can never finish in 1 second) but the same applies for
recipes which ran for less than an hour. A duration like "00:15:13" is
less ambiguous than "15:13".

Change-Id: Ic3253e8dd6ff268227d46567defd1639434bf4ad