Set systems to broken when de-associating with LC
If lab controllers are removed, attached systems should not remain in
'Automated' since it will lead to confusing test errors. This patch
marks them as broken so the schedule will not consider them for use.
Additionally, we validate before setting the system to automated, that
the a lab controller is associated.

Bug: 1347239
Change-Id: I6c08e529a6ef9b5d01d13511f8c56eb83b4fb5d5