test_database_migration: avoid leaking database connections
Strictly speaking, it's beaker-init which is leaking them, although for
the real beaker-init command it doesn't really matter because the
connections are released when the process exits.

In the tests, however, connections could be left lingering with open
transactions when the next test is trying to drop the database, which
would cause a deadlock. So let's close them properly.

Bug: 1356852
Change-Id: I3224724963cf4daf670a50a0b867801b568c8c1c