move online data migrations from beaker-init to beakerd
Instead of being a manual step which the admin must explicitly invoke
after their outage is finished, beakerd will automatically perform the
migrations. They are done as part of each iteration of the main
scheduler loop, so that it's not running concurrent transactions that
will fight for locks.

Completed migrations are recorded in the database (similar to how
Alembic tracks revisions) so that beakerd knows not to attempt them
again.

Bug: 1362414
Change-Id: I3810336d851606a8747db3a7e45df1380acaa306