recipe page: use bootstrap-collapse for task icons
This makes it simpler for us to ensure the icons are showing in the
right state, including when the page is newly opened.

The icons become real hyperlinks, although styled to keep them as plain
icons.

Requires nodejs >= 1.7 due to the reference to .fa-lg, which is defined
in FontAwesome with a variable substitution in its name. Older less is
unable to handle that correctly.

Based on a patch contributed by Matt Jia <mjia@redhat.com>.

Bug: 1358557
Change-Id: I8bdaa1bd67f3ddaaaecc8ad141ce9cfad2a3b7b9