Reserve system if recipe is in waiting state
If a user has scheduled short running tasks and the system is on high
load, beakerd might not be able to set the recipe to running after a
successful installation before tasks have completed. This results for
the recipe finishing as completed without reserving the system.

This patch includes the possibility that a recipe could be in a waiting
state and therefore honours the reservation request.

Bug: 1375035
Change-Id: I00efda65444d7fa091c4741e6cdd63faefb16f1b