tests: fix setup for test_GET_guests_host
The intention was to have the recipe running on the same LC as the
beaker-proxy process we are talking to: that is, self.get_lc(). But due
to a mix-up (self.system never passed into create_recipe) it was
actually creating a new system on a different LC.

It made no difference to the test because beaker-proxy will happily
serve up XML for *any* recipe, even recipes running on different lab
controllers. But let's fix the setup to match the original intention
(which is more realistic too).

I noticed this while debugging rogue commands in:
https://bugzilla.redhat.com/show_bug.cgi?id=1336272

Change-Id: Ia66675d1006103fbb4cec687fd8206299d85c8ea