set column defaults in __init__
In theory we should only need to specify the default value for
a particular column / ORM attribute in one place: the column definition
in our model.

However that has the downside that the default will not be populated on
newly constructed objects until after they are flushed. This is
a consequence of the fact that the defaults are at the SQL expression
level, which means they are only applied when the INSERT statement is
prepared and executed.

As a consequence, we end up either repeating the defaults as argument
defaults in __init__, or else we end up with inconsistent views of
a newly created object depending on whether we are looking at it before
or after it was flushed.

I hit this problem in commit ee11fa2e, when eliminating the duplicated
86400 reservesys duration. That commit added some logic to
RecipeReservationRequest.__init__ to apply the default, so that it would
still appear on newly created objects. But we can generalise it to apply
all defaults in MappedObject.__init__.

Doing this revealed a few oddities with RetentionTag and Power which are
tidied up in this patch as well.

Change-Id: I9c059c626375838317e0f1ff681b976e38942de6