beaker-repo-update: don't spew pointless warnings about RHEL repos
If you run beaker-repo-update without root access (as we do in our test
suite) on a RHEL system using subscription-manager, some code inside yum
will spew warnings to stderr like this:

    Repo rhel-6-server-rpms forced skip_if_unavailable=True due to: /etc/pki/entitlement/651570941082585099-key.pem

because the SSL client certificate is not readable. The warnings are not
relevant to beaker-repo-update, and we can silence them with only
a modest amount of hackery.

Change-Id: Ib8a6e694f95c541d1764b636a0c3688c65a2f6e4