beaker-create-kickstart: fix unicode warnings
It was spewing the usual SQLAlchemy warning to stderr:

    /usr/lib64/python2.6/site-packages/sqlalchemy/engine/default.py:459: SAWarning: Unicode type received non-unicode bind param value
      param.append(processors[key](compiled_params[key]))

because we were passing in command line arguments (strings) directly to
queries (which expect unicode).

I noticed this while updating the tests to invoke the command in
a subprocess, so now the tests will fail on any unexpected stderr spew
like this.

Change-Id: Id4620fe8c152383faed5f90e771cd83bf720449e