set INFO level for all our libraries' logging
Now that we have recently turned on OpenStack integration, there are
a number of new libraries whose debug spew is filling up the beakerd
log. Instead of continuing to turn off each library individual which is
spamming us, let's just make the default on the root logger INFO, and
set DEBUG for everything under 'bkr' which is application code.

Change-Id: Ibfd5c1af30d886d33a12ad89552aa857c5b2b2fd