tests: use run_createrepo() instead of invoking createrepo directly
Commit 56230eab moved the tests for beaker-repo-update from the Server
unit tests into IntegrationTests, but IntegrationTests doesn't have
a requirement on createrepo (only a transitive requirement on
createrepo_c from the server).

There is still a BuildRequires on createrepo, for the unit tests which
run during %check, which is why we didn't notice this problem straight
away. But with the change to make dogfood no longer do a build from git,
this missing dependency has become a problem.

There's really no reason we actually need createrepo specifically in
this test, so let's just use the run_createrepo() helper function which
is used by the server (and hence the dependencies are already declared).

Change-Id: I1a6f2d6444c5130c55bf6506c15165fd474523d1