tidy up setup.py files to make them suitable for uploading to PyPI
Changed the Python package names to match the RPM package names (people
will expect to run `pip install beaker-client`, not `pip install
bkr.client`).

Updated other metadata to be consistent across all packages, and re-use
descriptions from the RPM Summary tags.

Free-form license field is removed in favour of a specific Trove
classifier, as per
https://docs.python.org/2/distutils/setupscript.html#additional-meta-data
note 6.

For now we will only be uploading beaker-common and beaker-client to
PyPI as the other packages are still not likely to work properly when
installed through pip.

Bug: 1278605
Change-Id: Ib3883510e99081ec6b008ecce72c7753e95df244