tests: avoid timing issues with bkr distros-list --limit=1
bkr distros-list orders the distros by date_created descending, but it's
likely that multiple test cases will run within the same second so there
will be several distros with the same date_created value. We can't
assume that the one which will be returned is the one we just created.

Change-Id: Ib098baedc424f2c9ed108fd363aa31be1c22560f