Shows user friendly error message
This patch changes how users see an error message when creating a trust
account. If the trust can not be created due to a mis-configured server
configuration, the user will receive an error message with a 500 HTTP
status code.
If the server configuration is valid, but the user credentials are
entered wrong, the user will be given a more user friendly message.

Bug: 1395155
Change-Id: Iedd25c5cecc39213d76467a1f7078439c6d6acfe