remove last vestiges of qpid support
For future message bus integration we will not be using qpid.messaging
libraries (most likely Proton instead). And the existing code suffered
a few problems which would need some TLC to make it usable:

https://bugzilla.redhat.com/show_bug.cgi?id=1034043
https://bugzilla.redhat.com/show_bug.cgi?id=1034045

It will be easier to start afresh. Or this can be resurrected from git
history if needed.

Change-Id: Ifec7ff60028e4667101d21c5ebcfe05294933216