tests: fix timing issue in test_quiescent_period_only_applies_between_power_commands
Currently this test is only allowing exactly one polling loop interval
for the commands to be completed, which is occasionally not enough time
(because the command has to also be executed as well).

This patch makes the timeouts less aggressive (we use 2 * the loop
interval in the other, similar test cases) while still keeping the
quiescent period longer than the timeout to ensure that the test is
covering the desired behaviour.

Change-Id: I1f0e5e49ef21d2cc28d0c1170a720f2ce5c116fb