also log *before* handling an XMLRPC request
Currently the only way we can see what was in an XMLRPC request is the
debug logging which Beaker produces. The Apache access logs are useless
for XMLRPC because every request is just POST /client/.

However if the request hits a MemoryError this typically breaks all the
error handling code paths as well, which means that the log message at
the end is never logged and so it's impossible to know which XMLRPC
request triggered the MemoryError.

If we also log the details of the XMLRPC request *before* we start
handling it we should be able to identify the requests even if they
subsequently hit MemoryError.

Change-Id: Ib57c4f5a217ce4b21212214ec4fe985088571b76