tests: ensure we start with a pristine tasks directory
When the test suite starts the database is empty (there are no task
records) so let's make the basepath.rpms directory match.

Currently, if there are some files left behind in the basepath.rpms
directory from a previous test run, it will cause some tests in
bkr.inttest.client.test_task_add to fail.

The tests in bkr.inttest.server.selenium.test_tasks were avoiding this
problem by unlinking the files as soon as they were uploaded, but that
is actually backwards. A more realistic scenario is that the files are
*not* on disk for a fresh Beaker with an empty database, and they *are*
left on disk after uploading.

Change-Id: I93490a46c23e3e9e46bfa2c8252448d331a47675